Various changes
Modify the PickEmulator function such that it can also return the emulator for a given game OR the path to the emulator itself.

Implement the VerifySettings function.

Modify the FindFile function such that it accounts for shorter file names.

Impement a FindFolder function to search for a folder and return its full path.